Fix GET /notes/{id} and DELETE /notes/{id}
The AWS s3 api will return an error code of 'NoSuchKey' when using
getObject and the key does not exist while headObject and headBucket
will return an error code of 'NotFound'. Also, if deleteObject is used
on a nonexistent key, no error will be thrown, so deleteNoteById must
first check that an object exists.